Rename to match style guide
