+++ packages/metal-custom-element/src/define_web_component.js
